Save data initialization. Created a save class with local fields name, level, and time. Started a header toBytes() implementation for later serialization.

+++ aphelion/core/src/com/ephemerality/aphelion/spawn/entities/nob/Chest.java
@@ -5,7 +5,7 @@ import com.ephemerality.aphelion.graphics.SpriteSheet;
 public class Chest extends Nob{
 
 	public Chest(float x, float y) {
-		super(x, y, 32, 32, true, SpriteSheet.wood_box_0, 10);
+		super(x, y, 32, 32, true, SpriteSheet.wood_box_0, (short) 10);
 	}
 
 }
